requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-environ
python-dotenv
Pillow
Whitenoise
drf-yasg